First pass at extracting a CityDataManager.
Still working on understanding how village doors from a city should be
added to the village data store. The door list that is generated within
the Building class doesn't add doors just to the city, it looks for a
building id near the current building's ID, and adds the door to that
list. However, only the City's ID is actually added to the village list.